chore: move custom hook
